+++ src/traccuracy/metrics/__init__.py
@@ -1,6 +1,7 @@
 from ._basic import BasicMetrics
 from ._ctc import AOGMMetrics, CTCMetrics
 from ._divisions import DivisionMetrics
+from ._results import Results
 from ._track_overlap import TrackOverlapMetrics
 
 __all__ = [
@@ -8,5 +9,6 @@ __all__ = [
     "BasicMetrics",
     "CTCMetrics",
     "DivisionMetrics",
+    "Results",
     "TrackOverlapMetrics",
 ]
